backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
PyMySQL==1.0.3
python-weather==2.0.7
currencyconverter==0.17.0
FlightRadarAPI==1.3.34
skyscanner==1.1.4
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.2.3
scikit-learn==1.6.1
joblib==1.3.2
tensorflow-cpu==2.19.0
keras-tuner>=1.1.3

